move 'flip' tests to separate file
